requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
fastapi==0.116.1
h11==0.16.0
idna==3.10
pillow==11.3.0
pydantic==2.11.7
pydantic_core==2.33.2
pyfingerprint==1.5
pyserial==3.5
python-dotenv==1.1.1
requests==2.32.4
RPi.GPIO==0.7.1
sniffio==1.3.1
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
